Don't depend on projectile, use Cask
Projectile was needed by all templates to derive the project's name,
which is not enough to warrant such a large dependency.

Add a utility section to ptemplate-templates.el, which now contains a function
to get the project's name. Use it instead of `projectile-project-name'.

Use Cask to fix flycheck errors in .ptemplate.el files.

Fixes #1.