Emacs cask plugin template: add GPLv3 license
The Emacs cask plugin template now includes a license boilerplate in the
main file and generates a COPYING.txt at the project root.

Other licenses don't make sense for Emacs plugins, since most of them
are GPLv3-licensed anyway, and something like MIT is pointless because
there is no way to run Elisp in a proprietary context.

For all other languages, license management is much more complicated,
because there are many of them with different boilerplates and various
styles (SPDX vs HEADER in some cases) and will need to be addressed with
the help of another yet-to-be-written license-management plugin.